pyproject.toml
[tool.poetry]
name = "maxkb"
version = "0.1.0"
description = "智能知识库问答系统"
authors = ["shaohuzhang1 <shaohu.zhang@fit2cloud.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
django = "4.2.18"
djangorestframework = "^3.15.2"
drf-yasg = "1.21.7"
django-filter = "23.2"
langchain-openai = "^0.3.0"
langchain-anthropic = "^0.3.0"
langchain-community = "^0.3.0"
langchain-deepseek = "^0.1.0"
langchain-google-genai = "^2.0.9"
langchain-mcp-adapters = "^0.0.5"
langchain-huggingface = "^0.1.2"
langchain-ollama = "^0.3.0"
langgraph = "^0.3.0"
mcp = "^1.4.1"
psycopg2-binary = "2.9.10"
jieba = "^0.42.1"
diskcache = "^5.6.3"
pillow = "^10.2.0"
filetype = "^1.2.0"
torch = "2.6.0"
sentence-transformers = "^4.0.2"
openai = "^1.13.3"
tiktoken = "^0.7.0"
qianfan = "^0.3.6.1"
pycryptodome = "^3.19.0"
beautifulsoup4 = "^4.12.2"
html2text = "^2024.2.26"
django-ipware = "^6.0.4"
django-apscheduler = "^0.6.2"
pymupdf = "1.24.9"
pypdf = "4.3.1"
rapidocr-onnxruntime = "1.3.24"
python-docx = "^1.1.0"
xlwt = "^1.3.0"
dashscope = "^1.17.0"
zhipuai = "^2.0.1"
httpx = "^0.27.0"
httpx-sse = "^0.4.0"
websockets = "^13.0"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
gunicorn = "^23.0.0"
python-daemon = "3.0.1"
boto3 = "^1.34.160"
tencentcloud-sdk-python = "^3.0.1209"
xinference-client = "^1.3.0"
psutil = "^6.0.0"
celery = { extras = ["sqlalchemy"], version = "^5.4.0" }
django-celery-beat = "^2.6.0"
celery-once = "^3.0.1"
anthropic = "^0.49.0"
pylint = "3.1.0"
pydub = "^0.25.1"
cffi = "^1.17.1"
pysilk = "^0.0.1"
django-db-connection-pool = "^1.2.5"
opencv-python-headless = "^4.11.0.86"
pymysql = "^1.1.1"
oracledb = "^2.5.0"
accelerate = "^1.6.0"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"